Added Coordinadora USA, Package Consolidation, Dimensional Weight, other fixes
Functionality changes:

- Added Coordinadora USA
- In order to fully support the Coordinadora USA calculation, two
additional features were added: Package Consolidation and Dimensional
Weight
- If the Package Consolidation checkbox is checked, it will factor the
number of consolidated trackings in the freight cost calculation (only
applies to Coordinadora USA for now since this is the only courier that
charges for each extra tracking)
- Package Consolidation is set to a minimum of 2 packages (since it
wouldn't matter otherwise) and a maximum of 20
- 4-72 freight cost calculation is disabled if Package Consolidation is
enabled (since 4-72 doesn't support consolidation). The cheapest Pedidos
a USA tier is also disabled in this case, for the same reason.
- Dimensional Weight is only used for the Coordinadora USA calculation
(since it's the only courier listed so far that requires the
calculation). Technically, 4-72 depends on Dimensional Weight as well,
but the current official 4-72 calculator ignores the dimensional weight.
- On the couriers that take Dimensional Weight into account, the
Dimensional Weight will be used as the weight in the calculation if it's
higher than the Shipping Weight.
- The Shipping Out of Bogota ("¿Fuera de Bogota?") option is now taken
into account for the freight cost calculation (currently only
Coordinadora USA uses it).

Backend changes:
- I'm currently in the process of implementing a fillTableVal function
that will fill row values and will account for special
(courier-dependent) scenarios based on an additional parameter. This is
not fully implemented so far.

Aesthetical changes:
- Fixed wrong checkbox width in CSS
- When Dimensional Weight is preferred, for that specific calculation, a
Title tooltip was added to the weight for that courier, explaining that
the weight corresponds to Dimensional Weight.
- Added Title tooltips for Package Consolidation and Dimensional Weight